MapGenerator
The new layout allows for different sized rooms. The color of the cave mesh is changed to easily see where the walls. The collision with walls need debugging, if the player moves fast the collider ignores the wall and goes trough it!
I recommend we connect the rooms in a minimum spanning tree(using prims algorithm or some other algorithm...) then we use DFS to find a good place to spawn the portal to the next level