Game state manager
The manager only saves the health and scrap(for now!), it can be expanded to save other objects as well! During my testing I modified the health and ammo scripts. They are now attached to the canvas instead of the player and take the info dirrectly from the PlayerControllerMapTut script.